requirements.txt
Pygments==2.0.2
blessings==1.6
bpython==0.14.2
curtsies==0.1.19
greenlet==0.4.7
icalendar==3.9.0
nltk==3.0.3
numpy==1.9.2
python-dateutil==2.4.2
pytz==2015.4
requests==2.7.0
six==1.9.0
wheel==0.24.0
